+++ reproduce/example.cpp
@@ -7,8 +7,8 @@
 
 #include "../libpouq/quantizer.hpp"
 
-static constexpr size_t Dim = 256;
-static constexpr size_t N = Dim * 1000;
+static constexpr int Dim = 256;
+static constexpr int N = Dim * 1000;
 
 int
 main() {
@@ -21,7 +21,7 @@ main() {
   float l = 255.0f, u = 0.0f;
 
 #pragma omp parallel for reduction(min : l) reduction(max : u)
-  for (size_t i = 0; i < N; ++i) {
+  for (int i = 0; i < N; ++i) {
     data[i] = dis(gen);
     l = std::min(l, data[i]);
     u = std::max(u, data[i]);
@@ -31,7 +31,7 @@ main() {
 
   float mse_t = 0.0f;
 #pragma omp parallel for reduction(+ : mse_t)
-  for (size_t i = 0; i < N; ++i) {
+  for (int i = 0; i < N; ++i) {
     float code = std::round((data[i] - l) / step_size);
     code = std::clamp(code, 0.0f, 15.0f);
     float q = code * step_size + l;
@@ -45,7 +45,7 @@ main() {
 
   float mse_p = 0.0f;
 #pragma omp parallel for reduction(+ : mse_p)
-  for (size_t i = 0; i < N; i += Dim) {
+  for (int i = 0; i < N; i += Dim) {
     mse_p += quantizer.l2distance(data.data() + i, i);
   }
   std::cout << "POUQ MSE: " << mse_p / N << std::endl;
